Run Docker containers in the background
Previously, ANCM would actually *start* the  Docker container. That's fine when there are no errors. But when there is an issue, the Docker output is lost which makes it very hard to troubleshoot.

It's also slow, particularly when pulling/building an image for the first time.

This changes the approach to:
- do the Docker build/run in a separate first step, and *exit* the batch if that fails (enabling troubleshooting)
- have ANCM just *attach* to the already-running container, which is much faster and reliable

Also improve cleanup at the end of the batch script; force container removal via `docker rm -f`, to avoid conflicting container names or ports.